Restructure distance sorting JSON.
This fixes issue #150, where the JSON is wrong for distance sorting.
The structure is a little non-standard, as it is wrapped in an object
with hardcoded _geo_distance as name.

Reaching through inner-outer and tupled structs is unfortunate, but
seems to be nessecary in order to keep the API the same and to
keep everything nicely contained inside a GeoDistance struct as seen
from the consumer or user.